Agregando microservicio completo de StockFinal y ajuste en pedido

+++ src/main/java/com/cerveza/cerveza/controller/PedidoController.java
@@ -70,7 +70,7 @@ public class PedidoController {
     }
 
     @GetMapping("/producto/{idProducto}")
-    public List<PedidoDTO> buscarPorProducto(@PathVariable Long idProducto) {
+    public List<PedidoDTO> buscarPorProducto(@PathVariable Integer idProducto) {
         return pedidoService.buscarPorProducto(idProducto);
     }
 }
